+++ DXBase/DXBase/sources/Field/TurnFloor.cpp
@@ -6,7 +6,7 @@ TurnFloor::TurnFloor(IWorld * world, Vector2 & position) :spriteID_(-1),
 		Vector2{ position.x - (CHIPSIZE),position.y },
 		Vector2{ position.x ,position.y - (CHIPSIZE) },
 		Vector2{ position.x - (CHIPSIZE),position.y - (CHIPSIZE) }))
-	, defaultPos_(position), moveVelocity(1), moveCount_(CHIPSIZE*TURNRANGE)
+	, defaultPos_(position), moveVelocity(1), moveCount_(static_cast<int>(CHIPSIZE*TURNRANGE))
 {
 }
 TurnFloor::TurnFloor(int spriteID,IWorld * world, Vector2 & position) :spriteID_(spriteID),
@@ -15,7 +15,7 @@ TurnFloor::TurnFloor(int spriteID,IWorld * world, Vector2 & position) :spriteID_
 		Vector2{ position.x - (CHIPSIZE),position.y },
 		Vector2{ position.x ,position.y - (CHIPSIZE) },
 		Vector2{ position.x - (CHIPSIZE),position.y - (CHIPSIZE) }))
-	, defaultPos_(position), moveVelocity(1), moveCount_(CHIPSIZE*TURNRANGE)
+	, defaultPos_(position), moveVelocity(1), moveCount_(static_cast<int>(CHIPSIZE*TURNRANGE))
 {
 }
 
@@ -26,7 +26,7 @@ TurnFloor::TurnFloor(std::shared_ptr<TurnFloor> chip, IWorld * world, Vector2 & 
 		Vector2{ position.x ,position.y - (CHIPSIZE) },
 		Vector2{ position.x - (CHIPSIZE),position.y - (CHIPSIZE) }
 		))
-	, defaultPos_(position), moveVelocity(1), moveCount_(CHIPSIZE*TURNRANGE)
+	, defaultPos_(position), moveVelocity(1), moveCount_(static_cast<int>(CHIPSIZE*TURNRANGE))
 {
 }
 
@@ -37,7 +37,7 @@ TurnFloor::TurnFloor(TurnFloor & chip, IWorld * world, Vector2 & position) :
 		Vector2{ position.x,position.y - (CHIPSIZE) },
 		Vector2{ position.x - (CHIPSIZE),position.y - (CHIPSIZE) }
 		))
-	, defaultPos_(position), moveVelocity(1), moveCount_(CHIPSIZE*TURNRANGE)
+	, defaultPos_(position), moveVelocity(1), moveCount_(static_cast<int>(CHIPSIZE*TURNRANGE))
 {
 }
 
@@ -56,7 +56,7 @@ void TurnFloor::onUpdate(float deltaTime)
 {
 	defaultPos_ = position_- (moveVelocity*CHIPSIZE*TURNRANGE)+Vector2(CHIPSIZE*2,CHIPSIZE*2);
 
-	moveCount_+=1*(ceilf(deltaTime));
+	moveCount_+=1* static_cast<int>((ceilf(deltaTime)));
 	if (moveCount_ % 360 == 0) {
 		moveCount_ = 0;
 	}
